This commit was generated by cvs2svn to compensate for changes in r14123, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@14124 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/unix/x11/renderxv.c
@@ -112,7 +112,7 @@ int find_yuv_port(Display* display, XvPortID* port, fourcc_t* format)
     format_list = XvListImageFormats(display, adaptor_info[i].base_id,
 				     &num_formats);
 
-    for (j = 0; j < sizeof(fourcc_list)/sizeof(*fourcc_list); j++) {
+    for (j = 0; j < (int)(sizeof(fourcc_list) / sizeof(*fourcc_list)); j++) {
       if (format->id && fourcc_list[j].id != format->id) {
 	continue;
       }
